- added asdocs to DataGrid - added support for Apache Flex 4.12 to DataGridSkin - updated dispatchErrorEvent in ClassAvailableInstance - added file type property to LoadFile - added loaderContents property in LoadFile - added loaderByteArray property to LoadFile - added loaderContentType in LoadFile - fixed data property is set to value from file.data and loaderByteArray is set to loaderInfo.bytes - updated docs on LoadFile to correct loader bytes vs file bytes - added asdocs to WPAttachmentService - added file reference to WPAttachmentService - added fileData, fileName, dataField, contentType properties to WPAttachmentService - listening and checking for more events in WPAttachmentService - added deleteAttachment in WPService - fixed correct path in posts controller URLs in WPService - fixed bug when using WP_HOST redirect after login in WPService - added support to update site path on logout in WPService - added headerDividerWeight style to PanelSkin - added asdocs to DisplayObjectUtils - added walkDownComponentTree method in DisplayObjectUtils - set show display object outlines to true by default in MiniInspector - added popup property input to change property or style on the fly in MiniInspector - added asdocs to XMLUtils - added getAttributeSafeString, decodeAttributeString in XMLUtils - added enabled in ComponentDefinition - added default styles to DataGrid

+++ MainLibrary/src/com/flexcapacitor/services/WPServiceBase.as
@@ -108,6 +108,8 @@ package com.flexcapacitor.services {
 		 * */
 		public var permalinkPath:String = "api/";
 
+		public var originalSite:String = "";
+
 		private var _site:String = "";
 
 		/**
@@ -160,7 +162,7 @@ package com.flexcapacitor.services {
 				value = path + value;
 			}
 			else {
-				//value = value.replace(/?/, "&"); // replace first ? since we will be using it
+				value = value.replace(/\?/, "&"); // replace first ? since we will be using it
 				value = "?json=" + value;
 			}
 			
@@ -323,6 +325,8 @@ package com.flexcapacitor.services {
 			var token:String;
 			var anotherCallToGo:Boolean;
 			var profile:Boolean;
+			var error:Boolean;
+			var tokenFound:Boolean;
 			
 			profile ? responseTime = getTimer() - time :-3;
 			
@@ -335,7 +339,6 @@ package com.flexcapacitor.services {
 				profile ? currentTime = getTimer():-0;
 				json = JSON.parse(result);
 				profile ? parseTime = getTimer()- currentTime:-(1);
-				
 			}
 			catch (e:Error) {
 				// Error #1132: Invalid JSON parse input.
@@ -355,35 +358,39 @@ package com.flexcapacitor.services {
 				return;
 			}
 		
-			if (json && json is Object && "nonce" in json) {
-				token = json.nonce;
-			}
-			else if (json==null) {
+			if (json==null) {
 				serviceEvent.message = "JSON object is null";
 			}
 			
-			if (json) {
-				serviceEvent.data = json;
+			// token call
+			if (createPending || updatePending || uploadPending || deletePending) {
+				if (json && json is Object && "nonce" in json) {
+					token = json.nonce;
+					tokenFound = true;
+				}
 			}
 			
+			serviceEvent.data = json;
+			
 			if (json && json is Object && "status" in json && json.status=="error") {
+				error = true;
 				serviceEvent.message = json.error;//"Update token error";
-				dispatchEvent(serviceEvent);
-				return;
+				//dispatchEvent(serviceEvent);
+				//return;
 			}
-			else if (updatePending) {
+			else if (updatePending && tokenFound) {
 				updatePost(token);
 				anotherCallToGo = true;
 			}
-			else if (createPending) {
+			else if (createPending && tokenFound) {
 				createPost(token);
 				anotherCallToGo = true;
 			}
-			else if (deletePending) {
+			else if (deletePending && tokenFound) {
 				deletePost(token);
 				anotherCallToGo = true;
 			}
-			else if (uploadPending) {
+			else if (uploadPending && tokenFound) {
 				uploadAttachment(token);
 				anotherCallToGo = true;
 			}
@@ -399,9 +406,32 @@ package com.flexcapacitor.services {
 						var possibleSite:String = sites[0].siteurl;
 						
 						if (updateSitePathOnLogin) {
+							originalSite = site;
 							site = sites[0].path;
 						}
 					}
+				}
+			}
+			
+			if (call==WPServiceEvent.LOGOUT_USER) {
+				/*if (json && json is Object && json.blogs) {
+					
+					for each (var blog:Object in json.blogs) {
+						sites.push(blog);
+					}
+					
+					if (sites.length>0) {
+						var possibleSite:String = sites[0].siteurl;
+						
+						if (updateSitePathOnLogin) {
+							originalSite = site;
+							site = sites[0].path;
+						}
+					}
+				}*/
+				
+				if (updateSitePathOnLogin && originalSite && site != originalSite) {
+					site = originalSite;
 				}
 			}
